refactor code and add new icons
