+++ lib/presentation/home/views/home_view.dart
@@ -1,4 +1,4 @@
-import 'package:bookly/presentation/home/views/widgets/home_view_body.dart';
+import 'package:bookly/presentation/home/views/widgets/home/home_view_body.dart';
 import 'package:flutter/material.dart';
 
 class HomeView extends StatelessWidget {
