+++ lib/presentation/home/views/widgets/home/home_view_body.dart
@@ -1,15 +1,12 @@
-import 'package:blur/blur.dart';
-import 'package:bookly/presentation/home/views/widgets/best_seller_listview_item.dart';
-import 'package:bookly/presentation/home/views/widgets/custom_appbar.dart';
-import 'package:bookly/presentation/home/views/widgets/feature_listview_item.dart';
-import 'package:bookly/presentation/resources/assets.dart';
+import 'package:bookly/presentation/home/views/widgets/home/best_seller_listview_item.dart';
+import 'package:bookly/presentation/home/views/widgets/home/custom_appbar.dart';
+import 'package:bookly/presentation/home/views/widgets/home/feature_listview_item.dart';
 import 'package:bookly/presentation/resources/strings.dart';
 import 'package:bookly/presentation/resources/styles.dart';
 
 import 'package:bookly/presentation/resources/values.dart';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 
 class HomeViewBody extends StatelessWidget {
   const HomeViewBody({super.key});
